backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
boto3==1.34.7
pynamodb==5.5.1
python-jose==3.3.0
aws-wsgi==0.2.7 